refactor app and home to put into place to pull user data
