+++ client/src/components/Home.jsx
@@ -1,14 +1,26 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { Routes, Route, Link, Navigate } from "react-router-dom";
 import axios from "axios";
 
 import Header from "./Header.jsx";
 import LeftComponent from "./LeftComponent/LeftComponent.jsx";
-import CenterComponent from "./CenterComponent.jsx";
+import CenterComponent from "./CenterComponent/CenterComponent.jsx";
 import RightComponent from "./RightComponent/RightComponent.jsx";
 import Footer from "./Footer.jsx";
 
-export default function Home ({ authStatus, authenticate }) {
+export default function Home ({ authStatus, authenticate, currentUser }) {
+
+
+  useEffect(() => {
+    if (!currentUser) return;
+    handleGetUserData();
+  }, [currentUser])
+
+  const handleGetUserData = () => {
+    // make get request and give username to the server
+    console.log(currentUser);
+  }
+
   return (
     <>
       {!authStatus && (<Navigate to="/login" replace={true}/>)}
